+++ R/difNLR-package.R
@@ -45,7 +45,7 @@
 #' Package: difNLR\cr
 #' Type: Package\cr
 #' Version: 1.3.3\cr
-#' Date: 2020-04-23\cr
+#' Date: 2020-04-29\cr
 #' Depends: R (>= 3.1)\cr
 #' Imports: CTT, ggplot2 (>= 2.2.1), grDevices, methods, msm, nnet, plyr, reshape2, stats, VGAM\cr
 #' Suggests: ShinyItemAnalysis\cr